consistency: rename stdlib::def to stdlib::module
